Restructure utils folders and remove unnecessary directories - Add script files (import_companies, sort_companies) to pipelines/utils - Add DataFile to new data.py within agtern/pipelines (can add additional data helper classes here) - Add Internship model to agtern/models.py (can add additional models here)
