+++ agtern/__init__.py
@@ -1,3 +1,3 @@
 from .gui import *
+from .models import *
 from .pipelines import *
-from .utils import *
