+++ agtern/pipelines/__init__.py
@@ -1,3 +1,4 @@
 from .backend import *
+from .data import *
 from .frontend import *
 from .utils import *
